Changed NavGrid to use 32 bit instead of 64 bit integers

+++ Abaddon/src/Scene/Scripts/Models/Navigation/OpenNodesHeap.cpp
@@ -11,7 +11,7 @@ namespace Navigation
 		return false;
 	}
 
-	bool OpenNodesHeap::ContainsNodeIndex(size_t aNodeIndex)
+	bool OpenNodesHeap::ContainsNodeIndex(int32_t aNodeIndex)
 	{
 		for (const ValuedNode& valuedNode : c)
 			if (valuedNode.myNodeIndex == aNodeIndex) return true;
